timetable is working ish
for some reason when i am editing the SASS to show data on the time
table its not working.
The time table is working to a certain extent. It is sending data to the
database, but it is retiving the same info for all fields. as for some
reason the info on the database is not being transmited to tbl_timetable
as an ID as i wanted it to in the begining. Thus it is taking the id of
the first field, displaying it on the other fields.