Removed to do list as it was going to be too complex
reverted back to the Notepad and timetable pages to avoid confusion, and
to keep the website working.

As when i tried to implement the To-Do page i was not getting any errors
and getting a blank white screen with no indication of what was
happening. Tried reinstalling XAMPP in orted to fix it as most articles
online mentioned that PHP5 might have stopped working.
another fix was to implement "ini_set('display_errors', 1);"

when in fact thios was persent in the INDEX.php file. meaning that i was
rather lost on what to do